Bump core-ktx to 1.5.0 and appcompat to 1.3.0 in `android/tutorial`
Summary:
This diff bumps core-ktx from 1.3.0 to 1.5.0 and appcompat from 1.1.0 to 1.3.0 inside `android/tutorial`
core-ktx 1.5.0 is the last version built with Kotlin 1.4.x and is the latest we can safely upgrade to.

To update to core-ktx 1.6.0 we would need to bump KGP on the overall project to 1.4.x or 1.5.x to fix the build
failures.

Reviewed By: passy

Differential Revision: D29762647

fbshipit-source-id: 7ad23b3cbebe14b36dedaa7f8aedf11eece8429d