Rename websocketClientFlipperConnection
Summary:
At the moment Flipper spawns a WebSocket server which can be used to debug Web applications in the Browser i.e. Kite.

This diff renames websocketClientFlipperConnection to BrowserClientFlipperConnection.

Renamed to make it clearer what type of connection it really is. WebSocket connections will no longer have the same meaning once RSocket is removed and WebSockets are used instead.

Reviewed By: jknoxville

Differential Revision: D29937656

fbshipit-source-id: 99d55fcfe256865add1ba3aada2c23f935f5cabe