WebSocket-based server, base class and browser
Summary:
Introduce a base class for WebSocket based servers and a Browser-based implementation which is in use by Kite.

The implementation for the Browser-based one is basically taken as is from ServerController but slightly adapted to match the existing interface.

As with the RSocket-based implementation, this diff doesn't put this implementation into use but is a good opportunity to revisit the existing implementation.

Reviewed By: passy

Differential Revision: D29985886

fbshipit-source-id: 32abba37ec31478b6497ef5cfe90bb9aedc282d3