Lint for british spelling
Summary:
I hate when I have to comment on a diff and tell somebody to please not use
Her Majesty's Spelling, so I'll let the computer do the shouting.

Mixing two ways of spelling the same thing just isn't fun. I had to
work with an Android library that insisted on spelling it `colour`,
leading to awkward code like `colour: COLORS.BLUE` which is just annoying
and hard to remember.

Reviewed By: lblasa

Differential Revision: D30015807

fbshipit-source-id: 9f913e72617301273dbe12c60b9cdba8cea05537