WebSocket server GK
Summary:
This change enables a WebSocket server if the user is on the 'flipper_websocket_server' GK.

I don't expect this change to make it as the final product of the entire solution. Instead, it can be used to quickly enable/disable WebSocket during development/testing without requiring any code changes.

Reviewed By: passy

Differential Revision: D30069694

fbshipit-source-id: 75daac75e6dad7f63cf194d8287971c33d3abb61