Fix local CSR lookup
Summary:
From what I can tell, `idb file pull` seems to behave slightly differently for simulators. When pulling from a remote simulator, it appears to first create a directory for the local destination causing the read in line 508 to as it is trying to read the *folder* `app.csr` not `app.csr/app.csr`.

By omitting the destination file name, we avoid this behaviour.

Reviewed By: fabiomassimo

Differential Revision: D30541778

fbshipit-source-id: 851398aed5fd350a68ba0686707ce0f0aa09672c