Add warnings on how to handle offline emulators
Summary:
My emulators stopped showing up, but I noticed they were still listed in adb devices as 'offline'. Seems to be a weird ADB issue, changing the developer settings on the emulator fixed that for me, so linked to the SO answer where I found the solution.

Changelog: Added warning about offline emulators in ADB

Reviewed By: nikoant

Differential Revision: D30957786

fbshipit-source-id: 5468a983b5364c0fe868610978c09382f6b43d91