Fix bug in iOSDevice detection if both physical and emulator are active
Summary:
When connecting both to an iOS emulator and physical device, the emulator devices would continuesly be replaced, and Flipper would print the warning

`Tried to replace still connected device XXX with a new instance`

Fixed the logical mistake that causes it. Not sure if this caused any actual bugs, but at least it was incorrect.

Reviewed By: lblasa

Differential Revision: D31015451

fbshipit-source-id: 32dd29043e9dc48357fdbf68cde930d3be11419a