Fixed issue where plugins weren't reset properly when reconnecting a device
Summary: If a device disconnects and reconnects while keeping the plugin open, a new pluginInstance is available, but the React component would not be remounted since both the old and new pluginInstance have the same identifier (device/client/plugin). This change makes sure that if a different pluginInstance arrives, the rendering will be reinitialised as well.

Reviewed By: lblasa

Differential Revision: D31018845

fbshipit-source-id: ac09e7ac519c7aebfb4f8dc0cc7c6cda6fbbf633