Migrate away from deprecated instruments utility (#2903)
Summary:
The "instruments" utility was deprecated in Xcode 12 and removed now from Xcode 13, causing all the iOS 15 / Xcode 13  builds to fail to work with Flipper.

See deprecation notice here: https://developer.apple.com/documentation/xcode-release-notes/xcode-12-release-notes

> The instruments command is now deprecated in favor of its replacement: xctrace. Use xctrace to record, import, and export data from Instruments .trace files. (36641078)

The corresponding bug was reported here:
https://github.com/facebook/flipper/issues/2896

That call should be replaced with `xcrun xctrace` run and that is what we are doing here.

IMPORTANT: There was a workaround on killing the stale `instruments` proceses and that utility function is removed as part of this PR, as `instruments` were slow and inefficient, marked for deprecation more than a year ago. `xctrace` should not have that kind of issue and it doesn't make sense to keep this workaround now.

## Changelog

The "instruments" call replaced with `xcrun xctrace` use, to align with deprecation introduced in Xcode 12 and offer Xcode 13 compatibility.

Pull Request resolved: https://github.com/facebook/flipper/pull/2903

Test Plan:
In order to test this we should run the Flipper desktop app and make sure it connects to the Simulator started out of Xcode 13
It fails to do so in `master` version right now, see https://github.com/facebook/flipper/issues/2896

Reviewed By: mweststrate

Differential Revision: D31150276

Pulled By: passy

fbshipit-source-id: 0b45ae23c15a8481a91b1effe814176b04adbf3e