Refresh on new plugins
Summary:
afcoplan found a bug in the deeplink opening flow: If a device is open but not the app that supports a supported plugin, the dialogue won't disappear once the app is connected.

That is because apps are initialised with an empty plugin list and we're doing a reference equality check to get notified of changes which won't get triggered when the deep `plugin` property is changed.

Instead, we additionally wait for the `plugin-change` event that's emitted by the client to refresh.

Differential Revision: D31861521

fbshipit-source-id: 88fc456553e218f9c8ea68435ff84bdae0d1b586