Add changelog button to welcome screen
Summary:
## Overall goal
Decouple top menu from electron

## Approach

We have two options to solve it:

1. Leave the menu as-is for electron. Add a new menu to the left nav for other target envs.
2. Add a new menu to the left nav for all target envs. Leave only electron-specific options in the electron top menu.

This stack of diffs (D32311662, D32316987, D32329804, TDB) implements the second option. The primary reason is to have unified user-experience. If our users on different platforms have the same experience, it is easier for us to propagate knowledge and debug issues.

## Implementation plan

1. Move changelog to the welcome screen (D32311662)
2. Move import/export functionality to a dedicated menu in the left nav (D32316987)
3. Move plugin actions button to the toolbox with screen recording buttons (D32329804)
4. Collapse some left rail items under a single menu (D32356865)
5. Clean up the rest of electron menu items (D32357953)

## Followup

Return to re-grouping of the left rail buttons and toggles after decapitation

Reviewed By: mweststrate

Differential Revision: D32311662

fbshipit-source-id: e2c4bc1a7bd2c167335fcd6d76276ae7c4947859