Potential fix for artefact upload issue
Summary:
I've been investigating the now persistent upload issue for our artifacts and it appears to come down to artifacts being uploaded and downloaded as folder structures while they used to be individual files.

Annoyingly, the `ghr` tool we use for uploading throws a cryptic and misleading error: https://github.com/facebook/flipper/runs/4314584032?check_suite_focus=true

There is some more control in the `v2` version of the actions but for now, this seems like the safer option to fix this problem.

{F682461661}

Reviewed By: lblasa

Differential Revision: D32649218

fbshipit-source-id: 47ebefc434cb3b928b82a9fe8da3002b7b8465b6