Skip embedded plugins docs from indexing
Summary:
We generate two types of pages for plugin docs currently:
1) Full page used on Flipper docs website
2) Page without nav/header/footer which is used to embed docs into Flipper

The problem is that both of them now indexed and can both appear in search result.
See e.g. search results for keyword "android" which contains a lot of such embedded pages just because they have their first sections with name "Android" and because of that indexed to the top although shouldn't appear in search results at all.

{F683602640}

This diff disables indexing for embedded pages as it should be only possible to open them in Flipper.

Reviewed By: passy

Differential Revision: D32727589

fbshipit-source-id: 2068b8d39f19004a449098652c4c0cc2c7272d40